Release v1.5.0: full API audit, breaking fixes, Java 21, payload docs
Audited the SDK file-by-file against https://api.assinafy.com.br/v1/docs and
verified every finding live against the sandbox API.

Breaking:
- Remove AssignmentResource.get() (undocumented route, 404s for every call;
  use signerSelf.getSign() / GET /sign instead).
- SignerSelfResource.uploadSignature() now returns void and surfaces envelope
  errors (POST /signature returns a JSON envelope, not bytes).

Fixed:
- resetExpiration accepts null to clear expiration (+ clearExpiration()).
- resetPassword token is optional per the docs.
- FieldResource.validate(null) no longer NPEs.
- Binary download errors now surface the API error message/body.
- Signer create duplicate recovery handles HTTP 400 (not just 409).

Added:
- DocumentDetails.current_signer/page_count/created_by.
- Signer.step/notified/notification_history + AssignmentSignerNotification.
- Opt-in 429/503 retry (maxRetries) + ApiException.getRetryAfterSeconds().
- Verb+path Javadoc on every public resource method.
- docs/EXAMPLES.md with full request/response JSON for every method.

Tooling:
- Java 21 bytecode; CI matrix on JDK 21 + 25; SHA-pinned actions; Dependabot;
  release.yml (GitHub Packages); .gitlab-ci.yml parity; Maven Wrapper;
  enforcer, reproducible build, sources/javadoc jars via -Prelease.

Tests: 120 unit + 16 live smoke (sandbox-default), all green on JDK 21 and 25.

+++ src/main/java/com/assinafy/sdk/models/DocumentPage.java
@@ -8,8 +8,8 @@ public final class DocumentPage {
 
     private String id;
     private int number;
-    private double height;
-    private double width;
+    private int height;
+    private int width;
 
     @JsonProperty("download_url")
     private String downloadUrl;
@@ -20,11 +20,13 @@ public final class DocumentPage {
     public int getNumber() { return number; }
     public void setNumber(int number) { this.number = number; }
 
-    public double getHeight() { return height; }
-    public void setHeight(double height) { this.height = height; }
+    /** Page height in pixels. The API returns an integer (e.g. {@code 1651}). */
+    public int getHeight() { return height; }
+    public void setHeight(int height) { this.height = height; }
 
-    public double getWidth() { return width; }
-    public void setWidth(double width) { this.width = width; }
+    /** Page width in pixels. The API returns an integer (e.g. {@code 1275}). */
+    public int getWidth() { return width; }
+    public void setWidth(int width) { this.width = width; }
 
     public String getDownloadUrl() { return downloadUrl; }
     public void setDownloadUrl(String downloadUrl) { this.downloadUrl = downloadUrl; }
